Relocate main move functions to the game rather than the board/hand
